+++ src/os/kernel/shell/logic/shell.c
@@ -72,7 +72,7 @@ static void shell_handle_input()
 	if(del_err_stream)
 		delete(err_stream);
 
-	for(size_t i = 0; i < argc; i++)
+	for(int i = 0; i < argc; i++)
 	{
 		kfree(args[i]);
 	}
